+++ cat-home/src/main/java/com/dianping/cat/report/page/matrix/service/MatrixReportService.java
@@ -113,7 +113,7 @@ public class MatrixReportService extends AbstractReportService<MatrixReport> {
 		}
 	}
 
-	private MatrixReport queryFromWeeklyBinary(int id, String domain) {
+	private MatrixReport queryFromWeeklyBinary(long id, String domain) {
 		WeeklyReportContent content = m_weeklyReportContentDao.findByPK(id);
 
 		if (content != null) {
